+++ Borkle/Borkle/BubbleCanvas.swift
@@ -96,23 +96,7 @@ class BubbleCanvas: NSView {
         addTrackingArea(trackingArea)
     }
 
-    override var nextResponder: NSResponder? {
-        didSet {
-            print("OOGLE")
-        }
-    }
-
-    func printResponderChain() {
-        var responder: NSResponder? = self
-        while responder != nil {
-            print(responder as Any)
-            responder = responder?.nextResponder
-        }
-    }
-
     override func draw(_ areaToDrawPlzKthx: CGRect) {
-        printResponderChain()
-
         BubbleCanvas.background.set()
         bounds.fill()
