+++ Borkle/Borkle/UIResponder+Utilities.swift
@@ -1,0 +1,11 @@
+import AppKit
+
+extension NSResponder {
+    func printResponderChain() {
+        var responder: NSResponder? = self
+        while responder != nil {
+            print(responder as Any)
+            responder = responder?.nextResponder
+        }
+    }
+}
